+++ core-web/apps/dotcms-ui/src/app/portlets/shared/dot-content-types-edit/components/fields/dot-content-type-fields-variables/dot-content-type-fields-variables.component.html
@@ -1,4 +1,4 @@
-<dot-key-value
+<dot-key-value-ng
     [variables]="fieldVariables"
     (deletedVariable)="deleteExistingVariable($event)"
     (savedVariable)="updateExistingVariable($event)"
