+++ core-web/libs/ui/src/index.ts
@@ -17,7 +17,7 @@ export * from './lib/components/dot-info-page/dot-info-page.component';
 export * from './lib/components/dot-form-dialog/dot-form-dialog.component';
 export * from './lib/components/dot-pages-favorite-page-empty-skeleton/dot-pages-favorite-page-empty-skeleton.component';
 export * from './lib/components/dot-workflow-actions/dot-workflow-actions.component';
-export * from './lib/components/dot-key-value/dot-key-value.component';
+export * from './lib/components/dot-key-value-ng/dot-key-value-ng.component';
 
 // Directives
 export * from './lib/dot-field-required/dot-field-required.directive';
