+++ core-web/libs/ui/src/lib/components/dot-key-value-ng/dot-key-value-ng.component.scss
