+++ core-web/libs/ui/src/lib/components/dot-key-value-ng/dot-key-value-table-input-row/dot-key-value-table-input-row.component.scss
