+++ core-web/libs/ui/src/lib/components/dot-key-value-ng/util/dot-key-value-util.ts
@@ -1,4 +1,4 @@
-import { DotKeyValue } from '../dot-key-value.component';
+import { DotKeyValue } from '../dot-key-value-ng.component';
 
 export class DotKeyValueUtil {
     /**
